5. Aprofundando em testes e documentação/requirements.txt
asgiref==3.3.1
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==22.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
sqlparse==0.4.1
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
